Blinking pattern made with 50% duty cycle
In the previous implementation the first event would stick to 1s period whatever the outcome of
the secret verification was. The changing parameter in that case was the period of the LED being ON,
hence it's duty cycle was variable (1% for secret verified, 50% for secret not verified).

With this change the duty cycle will always be 50%, but the blinking frequency will be different
for  the verification outcome:
    T=1s    for secret not verified
    T=200ms for secret verified