README: Fix secret reading description
We previously described an interrupt service routing reading the secret
from a leaked secure box context. However, as the hello world application
is currently written, box 0 (not in an interrupt service routine) attempts
to read the secret from box 1 (without going through the secure gateway).
Update the README to reflect the reality of the current application.