Initialize the secret before handling button presses
Previously, we would initialize the secret lazily, on every call to
`challenge_verify`. We would also, in `challenge_verify`, communicate to
the outside world the pointer to the secret via `g_box_context`. Since we
would only verify the secret after one second, a user could press the
button before the secret pointer was communicated to the outside world.
This meant that the button handler would be able to successfully read from
`g_box_context`, which was at its default value of 0. This led to a
misleading print message, as the secret wasn't actually read from (instead,
address 0 was read from).

```
attempting to read the secret... done
```

Introduce `challenge_init`, a function to initialize the challenge, and
call it from `app_start` before registering the button handler. This
ensures that pressing the button will always attempt to read from the
proper challenge location.